Modularised the code by moving mock models,configs and env to separate files in test/utils
